+++ src/test/java/org/csanchez/jenkins/plugins/kubernetes/pipeline/KubernetesPipelineTest.java
@@ -24,128 +24,39 @@
 
 package org.csanchez.jenkins.plugins.kubernetes.pipeline;
 
-import static java.util.Arrays.*;
-import static org.csanchez.jenkins.plugins.kubernetes.KubernetesTestUtil.*;
 import static org.junit.Assert.*;
 
-import java.net.InetAddress;
-import java.net.URL;
 import java.util.Collections;
-import java.util.logging.Level;
 
-import org.apache.commons.compress.utils.IOUtils;
-import org.csanchez.jenkins.plugins.kubernetes.ContainerEnvVar;
-import org.csanchez.jenkins.plugins.kubernetes.ContainerTemplate;
 import org.csanchez.jenkins.plugins.kubernetes.KubernetesCloud;
-import org.csanchez.jenkins.plugins.kubernetes.PodTemplate;
-import org.csanchez.jenkins.plugins.kubernetes.model.KeyValueEnvVar;
-import org.csanchez.jenkins.plugins.kubernetes.model.SecretEnvVar;
-import org.csanchez.jenkins.plugins.kubernetes.model.TemplateEnvVar;
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.jenkinsci.plugins.workflow.job.WorkflowRun;
 import org.jenkinsci.plugins.workflow.test.steps.SemaphoreStep;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.junit.runners.model.Statement;
-import org.jvnet.hudson.test.BuildWatcher;
 import org.jvnet.hudson.test.Issue;
 import org.jvnet.hudson.test.JenkinsRuleNonLocalhost;
-import org.jvnet.hudson.test.LoggerRule;
-import org.jvnet.hudson.test.RestartableJenkinsRule;
-
-import com.google.common.collect.ImmutableMap;
 
 import hudson.model.Node;
 import hudson.slaves.DumbSlave;
 import hudson.slaves.NodeProperty;
 import hudson.slaves.RetentionStrategy;
 import io.fabric8.kubernetes.api.model.NamespaceBuilder;
-import io.fabric8.kubernetes.api.model.Secret;
-import io.fabric8.kubernetes.api.model.SecretBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
-import jenkins.model.JenkinsLocationConfiguration;
+import org.jvnet.hudson.test.RestartableJenkinsRule;
 
 /**
  * @author Carlos Sanchez
  */
-public class KubernetesPipelineTest {
-
-    private static final String SECRET_KEY = "password";
-    private static final String CONTAINER_ENV_VAR_VALUE = "container-env-var-value";
-    private static final String POD_ENV_VAR_VALUE = "pod-env-var-value";
-    private static final String CONTAINER_ENV_VAR_FROM_SECRET_VALUE = "container-pa55w0rd";
-    private static final String POD_ENV_VAR_FROM_SECRET_VALUE = "pod-pa55w0rd";
-
-    @ClassRule
-    public static BuildWatcher buildWatcher = new BuildWatcher();
+public class KubernetesPipelineTest extends AbstractKubernetesPipelineTest {
 
     @Rule
     public RestartableJenkinsRule story = new RestartableJenkinsRule();
     @Rule
-    public JenkinsRuleNonLocalhost r = new JenkinsRuleNonLocalhost();
-    @Rule
-    public LoggerRule logs = new LoggerRule().record(KubernetesCloud.class, Level.ALL);
-
-    @Rule
     public TemporaryFolder tmp = new TemporaryFolder();
-
-    private static KubernetesCloud cloud;
-
-    @BeforeClass
-    public static void configureCloud() throws Exception {
-        cloud = setupCloud();
-        createSecret(cloud.connect());
-    }
-
-    @Before
-    public void configureTemplates() throws Exception {
-        cloud.getTemplates().clear();
-        cloud.addTemplate(buildBusyboxTemplate("busybox"));
-        deletePods(cloud.connect(), Collections.emptyMap(), false);
-    }
-
-    @After
-    public void cleanup() throws Exception {
-        assertFalse("There are pods leftover after test execution, see previous logs",
-                deletePods(cloud.connect(), KubernetesCloud.DEFAULT_POD_LABELS, true));
-    }
-
-    /**
-     * Create a busybox template
-     */
-    private PodTemplate buildBusyboxTemplate(String label) {
-        // Create a busybox template
-        PodTemplate podTemplate = new PodTemplate();
-        podTemplate.setLabel(label);
-
-        ContainerTemplate containerTemplate = new ContainerTemplate("busybox", "busybox", "cat", "");
-        containerTemplate.setTtyEnabled(true);
-        podTemplate.getContainers().add(containerTemplate);
-        setEnvVariables(podTemplate);
-        return podTemplate;
-    }
-
-    @Before
-    public void addCloudToJenkins() throws Exception {
-        // Slaves running in Kubernetes (minikube) need to connect to this server, so localhost does not work
-        URL url = r.getURL();
-
-        String hostAddress = System.getProperty("jenkins.host.address");
-        if (hostAddress == null) {
-            hostAddress = InetAddress.getLocalHost().getHostAddress();
-        }
-        URL nonLocalhostUrl = new URL(url.getProtocol(), hostAddress, url.getPort(),
-                url.getFile());
-        JenkinsLocationConfiguration.get().setUrl(nonLocalhostUrl.toString());
-
-        r.jenkins.clouds.add(cloud);
-    }
 
     @Test
     public void runInPod() throws Exception {
@@ -337,48 +248,5 @@ public class KubernetesPipelineTest {
         assertNotNull(b);
         r.assertBuildStatusSuccess(r.waitForCompletion(b));
         r.assertLogContains("Apache Maven 3.3.9", b);
-    }
-
-    @Issue("JENKINS-46085")
-    @Test
-    public void containerLogs() throws Exception {
-        WorkflowJob p = r.jenkins.createProject(WorkflowJob.class, "containerLog");
-        p.setDefinition(new CpsFlowDefinition(loadPipelineScript("getContainerLog.groovy"), true));
-        WorkflowRun b = p.scheduleBuild2(0).waitForStart();
-        assertNotNull(b);
-        r.assertBuildStatusSuccess(r.waitForCompletion(b));
-        r.assertLogContains("INFO: Handshaking", b);
-        r.assertLogContains("INFO: Connected", b);
-    }
-
-    private String loadPipelineScript(String name) {
-        try {
-            return new String(IOUtils.toByteArray(getClass().getResourceAsStream(name)));
-        } catch (Throwable t) {
-            throw new RuntimeException("Could not read resource:[" + name + "].");
-        }
-    }
-
-    private static void createSecret(KubernetesClient client) {
-        Secret secret = new SecretBuilder()
-                .withStringData(ImmutableMap.of(SECRET_KEY, CONTAINER_ENV_VAR_FROM_SECRET_VALUE)).withNewMetadata()
-                .withName("container-secret").endMetadata().build();
-        client.secrets().createOrReplace(secret);
-        secret = new SecretBuilder().withStringData(ImmutableMap.of(SECRET_KEY, POD_ENV_VAR_FROM_SECRET_VALUE))
-                .withNewMetadata().withName("pod-secret").endMetadata().build();
-        client.secrets().createOrReplace(secret);
-    }
-
-    private static void setEnvVariables(PodTemplate podTemplate) {
-        TemplateEnvVar podSecretEnvVar = new SecretEnvVar("POD_ENV_VAR_FROM_SECRET", "pod-secret", SECRET_KEY);
-        TemplateEnvVar podSimpleEnvVar = new KeyValueEnvVar("POD_ENV_VAR", POD_ENV_VAR_VALUE);
-        podTemplate.setEnvVars(asList(podSecretEnvVar, podSimpleEnvVar));
-        TemplateEnvVar containerEnvVariable = new KeyValueEnvVar("CONTAINER_ENV_VAR", CONTAINER_ENV_VAR_VALUE);
-        TemplateEnvVar containerEnvVariableLegacy = new ContainerEnvVar("CONTAINER_ENV_VAR_LEGACY",
-                CONTAINER_ENV_VAR_VALUE);
-        TemplateEnvVar containerSecretEnvVariable = new SecretEnvVar("CONTAINER_ENV_VAR_FROM_SECRET",
-                "container-secret", SECRET_KEY);
-        podTemplate.getContainers().get(0)
-                .setEnvVars(asList(containerEnvVariable, containerEnvVariableLegacy, containerSecretEnvVariable));
     }
 }
